+++ core/src/main/java/io/github/CosecSecCot/Sprites/KingPig.java
@@ -6,7 +6,7 @@ import com.badlogic.gdx.physics.box2d.World;
 import io.github.CosecSecCot.Core;
 
 public class KingPig extends Pig {
-    public KingPig(World world, Core game, int x_pos, int y_pos) {
+    public KingPig(World world, Core game, float x_pos, float y_pos) {
         super(world, game, "bird_and_pigs", x_pos, y_pos, 41, 2, 126, 152, 0, -20, 12);
         this.points = 1000;
     }
