+++ core/src/main/java/io/github/CosecSecCot/Sprites/NormalPig.java
@@ -4,7 +4,7 @@ import com.badlogic.gdx.physics.box2d.*;
 import io.github.CosecSecCot.Core;
 
 public class NormalPig extends Pig {
-    public NormalPig(World world, Core game, int x_pos, int y_pos) {
+    public NormalPig(World world, Core game, float x_pos, float y_pos) {
         super(world, game, "bird_and_pigs", x_pos, y_pos, 732, 855, 48, 48, 0, -2, 10);
         this.points = 100;
     }
