+++ core/src/main/java/io/github/CosecSecCot/Sprites/Pig.java
@@ -6,7 +6,7 @@ import io.github.CosecSecCot.Core;
 public abstract class Pig extends Entity {
     protected int points;
 
-    public Pig(World world, Core game, String sprite_region, int x_pos, int y_pos, int sprite_x_pos, int sprite_y_pos, int width, int height, float xOffset, float yOffset, double health) {
+    public Pig(World world, Core game, String sprite_region, float x_pos, float y_pos, int sprite_x_pos, int sprite_y_pos, int width, int height, float xOffset, float yOffset, double health) {
         super(world, game, sprite_region, x_pos, y_pos, sprite_x_pos, sprite_y_pos, width, height, xOffset, yOffset, health);
         body.setUserData(this);
     }
